requirements.txt
asgiref==3.7.2
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.7.0
pytz==2024.1
sqlparse==0.4.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1